+++ recipes/default.rb
@@ -1,31 +1,23 @@
 #
-# Cookbook Name:: chef-orchestrator
-# Recipe:: default
+# Author:: Silvia Botros (<silvia.botros@sendgrid.com>)
+# Cookbook Name:: orchestrator
 #
-# Copyright (C) 2014 
+# Copyright (c) 2014
 #
-# All rights reserved - Do Not Redistribute
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
 #
-include_recipe "orchestrator::_database_setup"
-case node['platform']
-  when "debian", "ubuntu"
-    execute "download the .deb file" do
-      command "wget -P /tmp https://github.com/outbrain/orchestrator/releases/download/v1.0/orchestrator_1.0_amd64.deb"
-      action :run
-    end
-    execute "install from .deb" do
-      command "dpkg -i orchestrator_1.0_amd64.deb "
-      cwd "/tmp"
-      action :run
-      not_if "dpkg -l orchestrator"
-    end
-  when "redhat", "centos", "fedora" 
-    execute "install from rpm" do
-      command "sudo rpm -i https://github.com/outbrain/orchestrator/releases/download/v1.0/orchestrator-1.0-1.x86_64.rpm"
-      action :run
-      not_if "rpm -qa | grep orchestrator-1.0-1"
-    end
-end
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+include_recipe 'orchestrator::_package'
 
 template "/etc/orchestrator.conf.json" do
   source 'orchestrator.json.conf.erb'
@@ -35,9 +27,6 @@ file "/usr/local/orchestrator/conf/orchestrator.conf.json" do
   action :delete
 end
 
-execute "start_orchestrator" do
-  command "cd /usr/local/orchestrator && ./orchestrator http > /var/log/orchestrator.log 2>&1 &"
-  not_if "ps aux | grep orchestrator | grep -v grep"
+service 'orchestrator' do
+  action :start
 end
-
-